Use split/join to replace Windows escaped CWD
The issue is that the snapshot uses tcompare, which escapes backslashes
like JSON, so if the cwd appears in a string in an object, it gets its \
chars escaped to \\, and the resulting replacement against process.cwd()
doesn't hit a match.

Not sure why this wasn't working with a regexp or string replace()
method, but split/join works and I can only wrestle with \-escaped
slashes for so long before just giving up and doing whatever seems to
work.